chore(ssi): remove auto instrumentation utils
The auto instrumentation utils file and associated test file have
functions that are only imported in one other file. This commit moves
those functions to the relevant caller functions and removes the utils.
